refactor: URLQuery protocol로 구현
