+++ BoxOffice/Controller/ViewController.swift
@@ -11,7 +11,4 @@ class ViewController: UIViewController {
     override func viewDidLoad() {
         super.viewDidLoad()
     }
-    
-    
 }
-
