+++ BoxOffice/Network/URL/Implementation/URLQuery.swift
@@ -1,0 +1,12 @@
+//
+//  URLQuery.swift
+//  BoxOffice
+//
+//  Created by 훈민트 on 2023/08/18.
+//
+
+import Foundation
+
+struct URLQuery: QueryItemProtocol {
+    var queryItems: [URLQueryItem] = []
+}
